+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - `TryFirst` extension for `IEnumerable`
 - `TryLast` extension for `IEnumerable`
 - `TryElementAt` extension for `IEnumerable`
+- `Choose` extension for `IEnumerable`
 - `Collect` extension for `IEnumerable`
 ## 2.0.1
 ### Fixed